requirements.txt
autopep8==1.6.0
bidict==0.21.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
dnspython==2.2.0
docopt==0.6.2
eventlet==0.33.0
Flask==2.0.3
Flask-SocketIO==4.3.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
pycodestyle==2.8.0
python-engineio==3.13.2
python-socketio==4.6.0
requests==2.27.1
six==1.16.0
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.3
yarg==0.1.9
azure
azure-cognitiveservices-vision-customvision
Pillow